fix feedback at quickview and detail, show stock and status

+++ website_sellFlower/src/main/java/org/example/website_sellflower/service/Impl/ProductServiceImpl.java
@@ -26,7 +26,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public Product updateProduct(Integer id, Product updateproduct) {
+    public Product updateProduct(Long id, Product updateproduct) {
         Optional<Product> existingProduct = productRepository.findById(id);
         if(existingProduct.isPresent()) {
             Product product = existingProduct.get();
@@ -43,7 +43,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public boolean deleteProduct(Integer id) {
+    public boolean deleteProduct(Long id) {
         if(productRepository.existsById(id)){
             productRepository.deleteById(id);
             return true;
@@ -53,7 +53,7 @@ public class ProductServiceImpl implements ProductService {
     // ← THÊM MỚI
     @Override
     public Product findProductById(Long id) {
-        return productRepository.findById(id.intValue()).orElse(null);
+        return productRepository.findById(id).orElse(null);
     }
 
     @Override
